fpgamanager: Add pl-partial-final dts support for dfx and csoc
1. Add pl-partial-final dts support for dfx and csoc
2. User can provide additional pl-partial-custom dtsi file
   from SRC_URI of firmware recipe to modify pl-partial dtsi
   nodes. This dtsi file will get included to pl-partial-final.dts
   and converted to pl-partial-final.dtbo file.
3. Move pl-custom defines for base or static firmware recipes from
   device-tree recipes to fpgamanager_common bbclass.
4. Remove machine override from bbclass, Recipes that inherit from
   this class need to use an appropriate machine override for
   COMPATIBLE_MACHINE to build successfully.
5. If xsa contains more than one PR dtsi files then thrown an error
   from fpgamanager_dtg_dfx bbclass as this is not a supported from
   yocto.

+++ classes/fpgamanager_dtg_dfx.bbclass
@@ -17,19 +17,49 @@ python (){
 
     d.setVar("RP_XSCTH_HDF",[a for a in d.getVar('SRC_URI').split() if '.xsa' in a][0].lstrip('file://'))
 
-    #optional inputs
+    # Optional inputs
     if '.dtsi' in d.getVar("SRC_URI"):
-        d.setVar("CUSTOMPLINCLUDE_PATH",os.path.dirname([a for a in d.getVar('SRC_URI').split() if '.dtsi' in a][0].lstrip('file://')))
+        d.setVar("PL_PARTIAL_CUSTOM_INCLUDE_PATH",os.path.dirname([a for a in d.getVar('SRC_URI').split() if '.dtsi' in a][0].lstrip('file://')))
     if 'accel.json' in d.getVar("SRC_URI"):
         d.setVar("JSON_PATH",os.path.dirname([a for a in d.getVar('SRC_URI').split() if 'accel.json' in a][0].lstrip('file://')))
 }
 
+do_configure:append () {
+    # fpgamanager_dtg_dfx bbclass doesn't support multiple PR in signal xsa.
+    # DTG will suffix RpRm name to pl-partial-custom dtsi file when xsa has
+    # more than one PR DTG will generate pl-partial-custom-$RpRm.dtsi for each
+    # PR. Since this bbclass supports only one PR per xsa, it will find and use
+    # the first pl-partial-custom-$RpRm.dtsi file we find.
+    if [ $(find "${XSCTH_WS}/${XSCTH_DT_PATH}" -iname 'pl-partial-custom-*.dtsi' | wc -l) -gt 1 ]; then
+        bbfatal "XSA contains more than one pl-partial-custom dtsi in ${XSCTH_WS}/${XSCTH_DT_PATH} which is not supported from this bbclass"
+    else
+        for dtsi in `find "${XSCTH_WS}/${XSCTH_DT_PATH}" -iname 'pl-partial-custom*.dtsi'`; do
+            bbnote "Found pl-partial-custom dtsi file: $dtsi"
+            if [ -f "$dtsi" ]; then
+                # By default YAML_PARTIAL_OVERLAY_CUSTOM_DTS is set from yocto
+                # In some use case user can unset this variable from local.conf
+                # In such case copy PL_PARTIAL_CUSTOM_INCLUDE_PATH dtsi file only if
+                # YAML_PARTIAL_OVERLAY_CUSTOM_DTS is set else ignore it
+                if [ -f ${WORKDIR}/${PL_PARTIAL_CUSTOM_INCLUDE_PATH}/*.dtsi ]; then
+                    cp ${WORKDIR}/${PL_PARTIAL_CUSTOM_INCLUDE_PATH}/*.dtsi $dtsi
+                fi
+                break
+            else
+                bbfatal "No pl-partial-custom dtsi in ${XSCTH_WS}/${XSCTH_DT_PATH}"
+            fi
+        done
+    fi
+}
+
 do_install() {
     install -d ${D}${nonarch_base_libdir}/firmware/xilinx/${RP_PATH}/${PN}/
-    if [ -f ${B}/pl-partial*.dtbo ]; then
+
+    if [ -f ${B}/pl-partial-final*.dtbo ] && [ -n "${YAML_PARTIAL_OVERLAY_CUSTOM_DTS}" ]; then
+        install -Dm 0644 ${B}/pl-partial-final*.dtbo ${D}${nonarch_base_libdir}/firmware/xilinx/${RP_PATH}/${PN}/${PN}.dtbo
+    elif [ -f ${B}/pl-partial*.dtbo ] && [ -z "${YAML_PARTIAL_OVERLAY_CUSTOM_DTS}" ]; then
         install -Dm 0644 ${B}/pl-partial*.dtbo ${D}${nonarch_base_libdir}/firmware/xilinx/${RP_PATH}/${PN}/${PN}.dtbo
     else
-        bbfatal "A partial dtbo ending with ${B}/pl-partial-<partial_design>_inst_<n>.dtbo expected but not found"
+        bbfatal "A partial dtbo ending with ${B}/pl-partial-final-<partial_design>_inst_<n>.dtbo expected but not found"
     fi
 
     if [ "${SOC_FAMILY}" != "versal" ]; then
